% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sqldk-package}
\alias{sqldk-package}
\title{The sqldk package}
\description{
sqldk is a package to run duckdb SQL statements on R data frames.
}
\details{
There is a single user accessible function, sqldk.  It is usually called 
with a single argument which is a character string or vector specifying 
SQL statement(s).

See the \url{https://duckdb.org/docs} for information on duckdb itself.
}
\examples{
DF <- data.frame(a = 1:3, b = 4:6)
sqldk("select a, b, a + b as c from DF where a > 1")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrothendieck/sqldk}
  \item \url{https://duckdb.org/docs}
  \item Report bugs at \url{https://github.com/ggrothendieck/sqldk/issues}
}

}
