% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqldk.R
\name{sqldk}
\alias{sqldk}
\title{Run duckdb SQL statement(s) against R data frames.}
\usage{
sqldk(x, verbose = FALSE, db = "", register = FALSE, envir = parent.frame())
}
\arguments{
\item{x}{character vector of sql statements.}

\item{verbose}{TRUE or FALSE.}

\item{db}{"" or ":memory:" mean use in memory database; otherwise, provide
the name of an existing directory to use as an on disk database.  It will
be created if it does not already exist.}

\item{register}{if TRUE then tables are registered (like views) rather than 
copied to data base.  Default is TRUE.}

\item{envir}{environment in which to look for data frames.}
}
\value{
Result of last sql statement in x. (If no sql statement specified 
 then if db is a previously opened duckdb connection object then close it 
 and if not then open a new connection and return it.)
Commonly x is specified and all other arguments are left at default.
See \url{https://cran.r-project.org/web/packages/sqldk/README} for more info.
}
\description{
Run duckdb SQL statement(s) against R data frames.
}
\examples{
sqldk("select * from iris limit 3")
}
